requirements.txt
diskcache==5.6.3
Jinja2==3.1.3
pylint==3.1.0
PyYAML==6.0.1
